Properly install i3 configuration

+++ install.sh
@@ -51,8 +51,10 @@ mkdir -p "$HOME"/.config
 mkdir -p "$HOME"/.config/i3
 mkdir -p "$HOME"/.config/i3status
 
-ln -s "$dir"/i3config "$HOME"/.config/i3/config
-ln -s "$dir"/i3status_config "$HOME"/.config/i3status/config
+if [ ! -e "$HOME"/.config/i3/config ]; then
+    ln -s "$dir"/i3config "$HOME"/.config/i3/config
+    ln -s "$dir"/i3status_config "$HOME"/.config/i3status/config
+fi
 
 # Create local binary folder
 if [ ! -e "$HOME/bin" ]; then
